==> 2_preliminary_analysis_global.ipynb <==
# Notebook 2 - preliminary analysis, constant lifetime

## Intro

This Jupyter notebook contains parts of modeling behind the publication: 
> Krych, K. & Pettersen, JB. (2024). Long-term lifetime trends of large appliances since the introduction in Norwegian households. Journal of Industrial Ecology. 

Here, we perform the preliminary analysis, in which we vary all the model parameters to identify the critical ones, i.e., we perform a global uncertainty and sensitivity analysis.

We assume that the lifetime scale parameter follows a *constant* function.

## Imports and simulation setup

import sys, os
import numpy as np
import pandas as pd
import pickle
import matplotlib.pyplot as plt
import scipy.stats
from scipy.optimize import least_squares
import SALib.sample.sobol
import SALib.analyze.sobol
import time
import math
import seaborn as sns

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), os.pardir, 'ODYM/odym/modules')))
import ODYM_Classes as msc # import the ODYM class file

TimeStart = 1940
TimeEnd = 2022
SampleSize = 2**5

problem = {}
param_values = {}
variables = {}
MyYears = list(range(TimeStart,TimeEnd+1))
variables_for_UA = ['I', 'k', 'P', 'PpD', 'C', 'k-cab', 'lambda-cab', 'SoCE']

# all values below are relative changes in parameters (percentage changes)
u_i = 0.05
u_k = 0.1   
u_p = 0.01     
u_ppd = 0.01    
u_lambda_cab = 0.2 
u_k_cab = 0.2 
u_soce = 0.2     
u_c = 0.01    

## Functions

import funct # most functions are imported from funct.py

def lt_par_to_scale(t, lt_par):
    [C0] = lt_par
    scale = np.ones_like(t)*C0
    return scale

def calculate_popd_from_stock(t, stock, dict_values):
    dwellings = dict_values['P']/dict_values['PpD']
    s = dict_values['C']
    sf = np.zeros((len(t), len(t)))
    for m in range(0, len(t)):  # cohort index
        sf[m::,m] = scipy.stats.weibull_min.sf(np.arange(0,len(t)-m), c=dict_values['k-cab'], loc = 0, scale=dict_values['lambda-cab'])

    # MFA calculations start (assuming sf[0] != 0 and no negative inflows)
    i = np.zeros(len(t))
    s_c = np.zeros((len(t), len(t)))
    i[0] = s[0] / sf[0, 0]
    s_c[:, 0] = i[0] * sf[:, 0]
    for m in range(1, len(t)):
        i[m] = (s[m] - s_c[m, :].sum()) / sf[m,m]
        s_c[m::, m] = i[m] * sf[m::, m]
    
    o_c = np.zeros_like(s_c)
    o_c[1::,:] = -1 * np.diff(s_c,n=1,axis=0)
    o_c[np.diag_indices(len(t))] = i - np.diag(s_c) # allow for outflow in year 0 already

    soce = dict_values['SoCE'] # share of cabins electrified
    soce[soce >1] = 1
    TimeStart = t[0]
    soce[:1960-TimeStart] = 0 
    soce[soce <0] = 0
    el_cabins = np.einsum('tc,c->t',s_c,soce)
    popd = stock/(dwellings+el_cabins)
    return popd

def perform_MFA(lt_par, dict_values):
    t = dict_values['t']
    scale = lt_par_to_scale(t, lt_par)
    # create a 2D array with shape parameters
    lt_shape = np.full((len(t),len(t)), dict_values['k'], dtype=float)
    lt_shape = np.tril(lt_shape, 0)
    # create a 2D array with scale parameters
    lt_scale = np.repeat(np.reshape(scale,(len(scale),1)), repeats=len(scale), axis=1)
    lt_scale = np.tril(lt_scale, 0)
    lt = {'Type': 'Weibull', 'Shape': lt_shape,'Scale': lt_scale}
    # __find_unique_lt__
    params = {k:v for k,v in lt.items() if k!= 'Type'}
    sets = np.concatenate([[p] for p in params.values()], axis=0) # stacks all the p parameters
    sets = sets.reshape(len(params),-1) # reshapes from 3D form (p,t,t) into 2D form (p,t*t)
    unique, inverse = np.unique(sets, return_inverse=True, axis=1)
    inverse = inverse.reshape(len(t),len(t)) # reshapes from 1D form (t*t) into 2D form (t,t)
    length = np.shape(unique)[1]
    unique = {k:unique[p,:] for p,k in enumerate(params.keys())}
    # __compute_hz__
    hz_unique = np.zeros((len(t),length))
    for i in range(length): # for each unique parameter set
        if unique['Scale'][i] != 0:
            sf = scipy.stats.weibull_min.sf(np.arange(len(t)), c=unique['Shape'][i], loc = 0, scale=unique['Scale'][i])
            hz_unique[0,i] = 1-sf[0]
            for m in range(len(t)-1): # for each age m
                if sf[m] != 0:
                    hz_unique[m+1,i] = (sf[m] - sf[m+1]) / sf[m]
                else:
                    hz_unique[m+1,i] = 1
    hz = np.zeros((len(t), len(t)))
    for cohort in range(len(t)): # for each cohort c
        for time in range(cohort,len(t)): # for each year t 
            hz[time,cohort] = hz_unique[time-cohort,inverse[time,cohort]]
    # compute_inflow_driven_model
    i = dict_values['I']
    s_c = np.zeros((len(t), len(t))) # stock composition per year
    o_c = np.zeros((len(t), len(t))) # outflow compositionO
    for time in range(len(t)): # for each year t
        if time>0: # the initial stock is assumed to be 0
            o_c[time,:time] = s_c[time-1,:time] * hz[time,:time] 
            # subtract outflows of cohorts <m from the previous stock 
            s_c[time,:time] = s_c[time-1,:time] - o_c[time,:time]
        # Add new cohort to stock
        s_c[time,time] = i[time]
        o_c[time,time] = s_c[time,time] * hz[time,time]
    # popd calculation
    popd = calculate_popd_from_stock(t, s_c.sum(axis=1), dict_values)
    return popd, i, o_c, s_c, hz

def save_MFA_results(Dyn_MFA_System, dict_values, d, r, lt_fit):
    popd, i, o_c, s_c, hz = perform_MFA(lt_fit, dict_values)
    Dyn_MFA_System.FlowDict['F_01'].Values[:,d,r,0] = i
    Dyn_MFA_System.FlowDict['F_10c'].Values[:,:,d,r,0] = o_c
    Dyn_MFA_System.ExtraDict['F_10'].Values[:,d,r] = np.sum(o_c, axis=1)
    Dyn_MFA_System.StockDict['S_1c'].Values[:,:,d,r,0] = s_c
    Dyn_MFA_System.ExtraDict['S_1'].Values[:,d,r] = np.sum(s_c, axis=1)
    Dyn_MFA_System.ExtraDict['POpD'].Values[:,d,r]= popd
    Dyn_MFA_System.ExtraDict['hz'].Values[:,:,d,r]= hz
    Dyn_MFA_System.ExtraDict['LT-l'].Values[:,d,r] = lt_fit[0]
    return Dyn_MFA_System

def errfunc_var(lt_par, xdata, ydata, dict_values_r):
    popd, i, o_c, s_c, hz = perform_MFA(lt_par, dict_values_r)
    return popd[xdata-TimeStart]-ydata

## Model definition

ModelClassification_app  = {} # Create dictionary of model classifications

ModelClassification_app['Time'] = msc.Classification(Name = 'Time', Dimension = 'Time', ID = 1, 
                                                 Items = MyYears)
ModelClassification_app['Cohort'] = msc.Classification(Name = 'Cohort', Dimension = 'Time', ID = 2, 
                                                 Items = MyYears)
ModelClassification_app['Element'] = msc.Classification(Name= 'Elements', Dimension = 'Element', ID = 3, 
                                                    Items = ['C'])
ModelClassification_app['Durable'] = msc.Classification(Name = 'Durable', Dimension = 'Durable', ID = 4, 
                                                    Items = ['fridges' ,'freezers','washing machines', 'tumble dryers', 'dishwashers', 'ovens'])
ModelClassification_app['Run'] = msc.Classification(Name = 'Model run', Dimension = 'Run', ID = 5, 
                                                    Items = [])
aspects = ['Time','Element','Cohort', 'Durable', 'Run']
letters = ['t','e', 'c','d', 'r']
IndexTable = pd.DataFrame({'Aspect'        : aspects, # 'Time' and 'Element' must be present!
                           'Description'   : ['Model aspect ' + asp for asp in aspects],
                           'Dimension'     : [asp if asp != 'Cohort' else 'Time' for asp in aspects], # 'Time' and 'Element' are also dimensions
                           'Classification': [ModelClassification_app[Aspect] for Aspect in aspects],
                           'IndexLetter'   : letters}) # Unique one letter (upper or lower case) indices to be used later for calculations.
IndexTable.set_index('Aspect', inplace = True) 
IndexTable['IndexSize'] = \
    pd.Series([len(IndexTable.Classification[i].Items) for i in range(0,len(IndexTable.IndexLetter))], index=IndexTable.index)

Dyn_MFA_System = msc.MFAsystem(Name = 'PreliminaryApplianceSystem', 
                      Geogr_Scope = 'Norway', 
                      Unit = '1', 
                      ProcessList = [], 
                      FlowDict = {}, 
                      StockDict = {},
                      ParameterDict = {}, 
                      Time_Start = int(min(ModelClassification_app['Time'].Items)), 
                      Time_End = int(max(ModelClassification_app['Time'].Items)), 
                      IndexTable = IndexTable, 
                      Elements = IndexTable.loc['Element'].Classification.Items) # Initialize MFA system

Dyn_MFA_System.ProcessList = [] # Start with empty process list, only process numbers (IDs) and names are needed.
Dyn_MFA_System.ProcessList.append(msc.Process(Name = 'Environment', ID   = 0))
Dyn_MFA_System.ProcessList.append(msc.Process(Name = 'Process 1'  , ID   = 1))

Dyn_MFA_System.ParameterDict = {
    'I': msc.Parameter(Name = 'Inflow', Indices = 'td', Values=None, Unit = '1/yr'),
    'P': msc.Parameter(Name = 'Population', Indices = 't', Values=None, Unit = 'people/yr'),
    'PpD': msc.Parameter(Name = 'Person per dwelling', Indices = 't', Values=None, Unit = 'people/(dwelling*yr)'),
    'C': msc.Parameter(Name = 'Cabins', Indices = 't', Values=None, Unit = 'cabins/yr'),
    'k': msc.Parameter(Name = 'Durable lifetime (shape)', Indices = 'd', Values=None, Unit = '1'),
    'lambda-cab': msc.Parameter(Name = 'Cabin lifetime (scale)', Indices = '', Values=None, Unit = '1'),
    'k-cab': msc.Parameter(Name = 'Cabin lifetime (shape)', Indices = '', Values=None, Unit = '1'),
    'SoCE': msc.Parameter(Name = 'Share of cabins electrified', Indices = 't', Values=None, Unit = '%')}

excel = os.path.abspath(os.path.join(os.getcwd(), 'data.xlsx'))
for k in Dyn_MFA_System.ParameterDict.keys():
    indices = Dyn_MFA_System.ParameterDict[k].Indices
    Dyn_MFA_System.ParameterDict[k].Values = funct.read_parameter_excel(IndexTable, indices, excel, sheet_name=k)


variables = {'names': [], 'bounds': [], 'dists': []}
for var in variables_for_UA:
    uncert = globals()['u_'+var.lower().replace('-','_')]
    variables['names'].append(var)
    variables['bounds'].append([0,uncert])
    variables['dists'].append('norm')

problem = {
        'num_vars': len(variables['names']),
        'names': variables['names'], 
        'bounds': variables['bounds'], 
        'dists': variables['dists']
        }
# for normal 'norm' distribution, bounds indicate mean and st.dev

# sample
param_values = SALib.sample.sobol.sample(problem, SampleSize)
no_of_runs = np.shape(param_values)[0]

ModelClassification_app['Run'].Items = list(range(no_of_runs))
IndexTable.Classification['Run'].Items = list(range(no_of_runs))
IndexTable.IndexSize.loc['Run'] = no_of_runs

# Define the four flows a,b,c,d of the system, and initialise their values:
Dyn_MFA_System.FlowDict['F_01'] = msc.Flow(Name = 'Input', P_Start = 0, P_End = 1, Indices = 't,d,r,e', Values=None)
Dyn_MFA_System.FlowDict['F_10c'] = msc.Flow(Name = 'Output', P_Start = 1, P_End = 0, Indices = 't,c,d,r,e', Values=None)
Dyn_MFA_System.StockDict['S_1c'] = msc.Stock(Name='In-use stock', P_Res=1, Type=0, Indices='t,c,d,r,e', Values=None)
Dyn_MFA_System.StockDict['dS_1'] = msc.Stock(Name='In-use stock change', P_Res=1, Type=1, Indices='t,d,r,e',Values=None)

Dyn_MFA_System.ExtraDict = {}
Dyn_MFA_System.ExtraDict['S_1'] = msc.Parameter(Name = 'In-use stock total', Indices = 'tdr', Values=None)
Dyn_MFA_System.ExtraDict['F_10'] = msc.Parameter(Name = 'Outflows total', Indices = 'tdr', Values=None)
Dyn_MFA_System.ExtraDict['POpD'] = msc.Parameter(Name = 'Product ownership per dwelling', Indices = 'tdr', Values=None) 
Dyn_MFA_System.ExtraDict['hz'] = msc.Parameter(Name = 'Hazard matrix', Indices = 'tcdr', Values=None)
Dyn_MFA_System.ExtraDict['LT-l'] = msc.Parameter(Name = 'Durable lifetime (scale)', Indices = 'tdr', Values=None)

for key in Dyn_MFA_System.ExtraDict.keys():
    Dyn_MFA_System.ExtraDict[key].Values = np.zeros(tuple([len(IndexTable.set_index('IndexLetter').loc[x]['Classification'].Items) for x in Dyn_MFA_System.ExtraDict[key].Indices]))
Dyn_MFA_System.Initialize_FlowValues()
Dyn_MFA_System.Initialize_StockValues() 
Dyn_MFA_System.Consistency_Check()

## Analysis

df_popd = funct.read_data_excel(IndexTable, 'td',excel, sheet_name='POpD_data')
start_regression = time.time()
for d,durable in enumerate(ModelClassification_app['Durable'].Items):
    start_regression_d = time.time()
    print(f"Fitting lifetime for {durable}...")
    dict_values = funct.reduce_dimensions(IndexTable, durable, Dyn_MFA_System.ParameterDict)
    df_d = df_popd[df_popd['durable'] == durable]
    xdata = df_d['time'].values
    ydata = df_d['value'].values
    for r in range(no_of_runs):
        dict_values_r = funct.multiply_by_uncertainty(durable, problem, param_values[r,:], dict_values,Dyn_MFA_System.ParameterDict)
        dict_values_r['t'] = MyYears
        lt_par0 = [20] 
        bounds = ([0],[40])
        lt_fit_result = least_squares(errfunc_var, lt_par0, bounds=bounds, args=(xdata, ydata, dict_values_r))
        
        Dyn_MFA_System = save_MFA_results(Dyn_MFA_System, dict_values_r, d, r, lt_fit_result.x)
        if r % 100 == 0 and r != 0:
            print(f'iteration {r} in {no_of_runs}')
        
    end_regression_d = time.time()
    print(f'Time elapsed for {durable}: {math.trunc((end_regression_d-start_regression_d)/60)} min {round((end_regression_d-start_regression_d) % 60)} s')
end_regression = time.time()
print(f'Time elapsed total: {math.trunc((end_regression-start_regression)/60)} min {round((end_regression-start_regression) % 60)} s')

# calculate mean lifetime
mu = np.zeros([len(ModelClassification_app['Time'].Items), len(ModelClassification_app['Durable'].Items)+1, len(ModelClassification_app['Run'].Items)])
for d,durable in enumerate(ModelClassification_app['Durable'].Items):
    dict_values = funct.reduce_dimensions(IndexTable, durable, Dyn_MFA_System.ParameterDict)
    for r in range(no_of_runs):
        dict_values_r = funct.multiply_by_uncertainty(durable, problem, param_values[r,:], dict_values,Dyn_MFA_System.ParameterDict)
        scale = Dyn_MFA_System.ExtraDict['LT-l'].Values[:,d,r]
        mu[:,d,r] = funct.mean_from_scale_parameters(scale, dict_values_r)
Dyn_MFA_System.ExtraDict['mu'] = msc.Parameter(Name = 'Durable lifetime (mean)', Indices = 'tdr', Values=mu)

# calculate age
age_o_i = np.empty((len(MyYears), len(ModelClassification_app['Durable'].Items), no_of_runs))
age_o = np.empty((len(MyYears), len(ModelClassification_app['Durable'].Items), no_of_runs))
for d, durable in enumerate(ModelClassification_app['Durable'].Items):
    for r in range(no_of_runs):
        o_c = Dyn_MFA_System.FlowDict['F_10c'].Values[:,:,d,r,0]
        age_o_i[:,d,r] = funct.calculate_age(o_c,isstock=False,inflows=Dyn_MFA_System.FlowDict['F_01'].Values[:,d,r,0])
        age_o[:,d,r] = funct.calculate_age(o_c,isstock=False)
Dyn_MFA_System.ExtraDict['age_o_i'] = msc.Parameter(Name = 'Mean age of outflows, scaled by inflows', Indices = 'tdr', Values=age_o_i)
Dyn_MFA_System.ExtraDict['age_o'] = msc.Parameter(Name = 'Mean age of outflows', Indices = 'tdr', Values=age_o)

with open("2_preliminary_analysis_global.dat", "wb") as f:
     pickle.dump([Dyn_MFA_System, problem, param_values, no_of_runs], f)

## Plotting

x = IndexTable.Classification['Time'].Items
columns = 2
rows = 4
fig = plt.figure(figsize=(6*columns, 3.5*rows), constrained_layout=True)
# fig.set_dpi(dpi) #min 600 for publication
gs = fig.add_gridspec(rows, columns,hspace=0.1)
row = 0
col = 0
for d,durable in enumerate(ModelClassification_app['Durable'].Items):
    ax = fig.add_subplot(gs[row, col])
    for r in range(no_of_runs):
        ax.plot(MyYears, mu[:,d,r], color='C0', alpha=0.2)
    ax.set_xlabel('Year')
    ax.set_ylabel('Mean lifetime (years)')
    ax.set_title(durable.capitalize())
    ax.set_ylim(8, 37)
    if col == columns-1:
        col = 0
        row+= 1
    else:
        col += 1

x = IndexTable.Classification['Time'].Items
columns = 2
rows = 4
fig = plt.figure(figsize=(6*columns, 3.5*rows), constrained_layout=True)
# fig.set_dpi(dpi) #min 600 for publication
gs = fig.add_gridspec(rows, columns,hspace=0.1)
row = 0
col = 0
df_popd = funct.read_data_excel(IndexTable, 'td',excel, sheet_name='POpD_data')
for d,durable in enumerate(ModelClassification_app['Durable'].Items):
    df_d = df_popd[df_popd['durable'] == durable]
    ax = fig.add_subplot(gs[row, col])
    for r in range(no_of_runs):
        popd_model = Dyn_MFA_System.ExtraDict['POpD'].Values[:,d,r]
        ax.plot(MyYears, popd_model, color='C0', alpha=0.1, zorder=r)
    ax.set_xlabel('Year')
    ax.set_ylabel('Product ownership per dwelling')
    ax.set_title(durable.capitalize())
    ax.scatter(df_d['time'].values,df_d['value'].values, color='black', zorder=r+1)
    if col == columns-1:
        col = 0
        row+= 1
    else:
        col += 1
plt.xlim(1940-4,2022+4)
plt.show()

sobol_St = {}
sobol_S1 = {}
sobol_St_CI = {}
sobol_S1_CI = {}
for d,durable in enumerate(IndexTable.Classification['Durable'].Items):
    problem_d = problem
    sobol_indices = [SALib.analyze.sobol.analyze(problem, Y) for Y in Dyn_MFA_System.ExtraDict['LT-l'].Values[:,d,:no_of_runs]]
    sobol_St[durable] = np.array([s['ST'] for s in sobol_indices])
    sobol_S1[durable] = np.array([s['S1'] for s in sobol_indices])
    sobol_St_CI[durable] = np.array([s['ST_conf'] for s in sobol_indices])
    sobol_S1_CI[durable] = np.array([s['S1_conf'] for s in sobol_indices])

S = sobol_St[durable]

data = {}
for i in range(0,len(problem["names"])):
    name = problem["names"][i]
    values = S[:, i][20:] # fix here!!! it only shows years 
    data[name] = values
df = pd.DataFrame(data)
df.index = IndexTable.Classification['Time'].Items[20:]
df.index.names = ['Year']

labels = problem["names"]
labels.reverse()
plt.stackplot(df.index,
            [df[i] for i in labels],
            labels=labels, colors=sns.color_palette("Spectral", len(problem["names"])))

plt.legend(reversed(plt.legend().legendHandles), reversed(labels), loc=2, fontsize='large', bbox_to_anchor=(1, 1))
plt.show()

sobol_St_scaled = {durable: sobol_St[durable][-1, :]/np.sum(sobol_St[durable][-1, :]) for durable in ModelClassification_app['Durable'].Items}

durables = ModelClassification_app['Durable'].Items
weight_counts = {name: np.array([sobol_St_scaled[durable][i] for durable in ModelClassification_app['Durable'].Items]) for i,name in enumerate(problem['names'])}

fig, ax = plt.subplots()
bottom = np.zeros(len(durables))
for name, weight_count in weight_counts.items():
    p = ax.bar(durables, weight_count, 0.5, label=name, bottom=bottom)
    bottom += weight_count

# ax.set_title("Number of penguins with above average body mass")
handles, labels = ax.get_legend_handles_labels()
plt.xticks(rotation=90)
ax.legend(handles[::-1], labels[::-1], loc="upper left", bbox_to_anchor=(1, 1))

plt.show()
